requirements.txt
python-gnupg==0.4.7
requests==2.26.0
